pyproject.toml
[tool.poetry]
name = "fundmgr"
version = "0.1.0"
description = ""
authors = ["Luke Roberts <contact@ltroberts.co.uk>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
django-filter = "^25.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.black]
line-length = 79
target-version = ['py312']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
